IRC class in a separate file, with comments
